refactor(verifier): centralize engine liveness helpers
