+++ rust/fsl-verifier/src/lib.rs
@@ -6,7 +6,9 @@ mod agreement;
 mod bmc;
 mod eval;
 mod induction;
+mod liveness;
 mod refinement;
+mod trace;
 mod transition;
 mod value;
